[HTML/CSS] Use Abbreviated Hex Code
- Use abbreviated hexadecimal values to set the color of the `body`
  element

  - Change the `background-color` of the `body` element from `#808080`
    to `#F00`

NOTES:

Many people feel overwhelmed by the possibilities of more than 16
million colors. And it's difficult to remember hex code. Fortunately,
you can shorten it.

For example, red, which is `#FF0000` in hex code, can be shortened to
`#F00`. That is, one digit for red, one digit for green, one digit for
blue.

This reduces the total number of possible colors to around 4,000. But
browsers will interpret `#FF0000` and `#F00` as exactly the same color.

  To be precise about the number of color combinations when using
  abbreviated hex color values:

    16^3 = 4096 -> 4,096 color combinations!